move default addresses to rethnet, not _rpc_server
addresses review comment https://github.com/NomicFoundation/hardhat/pull/4057#discussion_r1257328177
