+++ crates/rethnet_rpc_server/src/lib.rs
@@ -52,30 +52,6 @@ struct AppState {
 
 type StateType = Arc<AppState>;
 
-pub const DEFAULT_ACCOUNTS: [&str; 20] = [
-    // these were taken from the standard output of a run of `hardhat node`
-    "f39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
-    "70997970C51812dc3A010C7d01b50e0d17dc79C8",
-    "3C44CdDdB6a900fa2b585dd299e03d12FA4293BC",
-    "90F79bf6EB2c4f870365E785982E1f101E93b906",
-    "15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
-    "9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
-    "976EA74026E726554dB657fA54763abd0C3a0aa9",
-    "14dC79964da2C08b23698B3D3cc7Ca32193d9955",
-    "23618e81E3f5cdF7f54C3d65f7FBc0aBf5B21E8f",
-    "a0Ee7A142d267C1f36714E4a8F75612F20a79720",
-    "Bcd4042DE499D14e55001CcbB24a551F3b954096",
-    "71bE63f3384f5fb98995898A86B02Fb2426c5788",
-    "FABB0ac9d68B0B445fB7357272Ff202C5651694a",
-    "1CBd3b2770909D4e10f157cABC84C7264073C9Ec",
-    "dF3e18d64BC6A983f673Ab319CCaE4f1a57C7097",
-    "cd3B766CCDd6AE721141F452C550Ca635964ce71",
-    "2546BcD3c84621e976D8185a91A922aE77ECEc30",
-    "bDA5747bFD65F08deb54cb465eB87D40e51B197E",
-    "dD2FD4581271e230360230F9337D5c0430Bf44C0",
-    "8626f6940E2eb28930eFb4CeF49B2d1F2C9C1199",
-];
-
 fn error_response_data<T>(msg: &str) -> ResponseData<T> {
     event!(Level::INFO, "{}", &msg);
     ResponseData::new_error(0, msg, None)
@@ -562,34 +538,8 @@ impl Server {
     /// accepts a configuration and a set of initial accounts to initialize the state.
     pub async fn new(
         config: Config,
-        genesis_accounts: Option<HashMap<Address, AccountInfo>>,
+        genesis_accounts: HashMap<Address, AccountInfo>,
     ) -> Result<Self, Error> {
-        let genesis_accounts = genesis_accounts.unwrap_or(DEFAULT_ACCOUNTS.iter().fold(
-            HashMap::default(),
-            |mut genesis_accounts, account| {
-                use std::str::FromStr;
-                genesis_accounts.insert(
-                    Address::from_str(account)
-                        .map_err(|e| {
-                            let e = Error::AddressParse {
-                                address: account.to_string(),
-                                reason: e.to_string(),
-                            };
-                            event!(Level::ERROR, "{e}");
-                            e
-                        })
-                        .unwrap(),
-                    AccountInfo {
-                        balance: U256::from(10000), // copied from Hardhat Network stdout
-                        nonce: 0,
-                        code: None,
-                        code_hash: KECCAK_EMPTY,
-                    },
-                );
-                genesis_accounts
-            },
-        ));
-
         let listener = TcpListener::bind(config.address).map_err(Error::Listen)?;
         event!(Level::INFO, "Listening on {}", config.address);
